requirements.txt
anyio==4.12.1
arrow==1.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastkml==1.4.0
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pygeoif==1.6.0
pyparsing==3.2.5
pyTelegramBotAPI==4.29.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.6
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
